requirements.txt
pandas==1.3.4
tqdm==4.62.3
pytorch==1.10.0
numpy==1.21.2
scikit-learn==0.23.2
cudatoolkit==10.2.89
python==3.9.7